[Screenshotting] Server package for stateless code (#188390)
## Summary

This PR moves code that is usually impacted by a Puppeteer/Chromium
upgrade to a package. The same work is happening to a lesser degree in
the 7.17 branch. Together, these changes will ease the manual labor of
backporting the upgrades from main to 7.17.

---------

(cherry picked from commit bf673d9c565c288223483c4c622f37f745ce70e6)

# Conflicts:
#	.github/CODEOWNERS
