+++ packages/kbn-screenshotting-server/README.md
@@ -1,0 +1,3 @@
+# @kbn/screenshotting-server
+
+Stateless code pertaining to the capture of screenshots for Kibana Reporting
