+++ packages/kbn-screenshotting-server/kibana.jsonc
@@ -1,0 +1,5 @@
+{
+  "type": "shared-server",
+  "id": "@kbn/screenshotting-server",
+  "owner": "@elastic/appex-sharedux"
+}
