+++ packages/kbn-screenshotting-server/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@kbn/screenshotting-server",
+  "private": true,
+  "version": "1.0.0",
+  "license": "SSPL-1.0 OR Elastic License 2.0"
+}
