+++ x-pack/plugins/screenshotting/server/browsers/chromium/driver.test.ts
@@ -7,9 +7,9 @@
 
 import type { Logger } from '@kbn/logging';
 import { ScreenshotModePluginSetup } from '@kbn/screenshot-mode-plugin/server';
+import { ConfigType } from '@kbn/screenshotting-server';
 import * as puppeteer from 'puppeteer';
 import { Size } from '../../../common/layout';
-import { ConfigType } from '../../config';
 import { PreserveLayout } from '../../layouts/preserve_layout';
 import { HeadlessChromiumDriver } from './driver';
