+++ x-pack/plugins/screenshotting/server/browsers/chromium/driver_factory/index.ts
@@ -7,23 +7,22 @@
 
 import type { Logger } from '@kbn/core/server';
 import type { ScreenshotModePluginSetup } from '@kbn/screenshot-mode-plugin/server';
+import { ConfigType, args } from '@kbn/screenshotting-server';
 import { getDataPath } from '@kbn/utils';
 import { spawn } from 'child_process';
 import del from 'del';
 import fs from 'fs';
 import { uniq } from 'lodash';
 import path from 'path';
-import puppeteer, { Browser, ConsoleMessage, Page, Viewport, PageEvents } from 'puppeteer';
+import puppeteer, { Browser, ConsoleMessage, Page, PageEvents, Viewport } from 'puppeteer';
 import { createInterface } from 'readline';
 import * as Rx from 'rxjs';
-import { catchError, concatMap, ignoreElements, mergeMap, map, reduce, takeUntil, tap } from 'rxjs';
-import { PerformanceMetrics } from '../../../../common/types';
+import { catchError, concatMap, ignoreElements, map, mergeMap, reduce, takeUntil, tap } from 'rxjs';
 import { getChromiumDisconnectedError } from '..';
 import { errors } from '../../../../common';
-import { ConfigType } from '../../../config';
+import { PerformanceMetrics } from '../../../../common/types';
 import { safeChildProcess } from '../../safe_child_process';
 import { HeadlessChromiumDriver } from '../driver';
-import { args } from './args';
 import { getMetrics } from './metrics';
 
 interface CreatePageOptions {
