+++ x-pack/plugins/screenshotting/server/browsers/download/index.test.ts
@@ -5,13 +5,13 @@
  * 2.0.
  */
 
-import path from 'path';
-import mockFs from 'mock-fs';
+import { ChromiumArchivePaths, PackageInfo } from '@kbn/screenshotting-server';
 import { access, readdir } from 'fs/promises';
-import { ChromiumArchivePaths, PackageInfo } from '../chromium';
-import { fetch } from './fetch';
-import { sha256 } from './checksum';
+import mockFs from 'mock-fs';
+import path from 'path';
 import { download } from '.';
+import { sha256 } from './checksum';
+import { fetch } from './fetch';
 
 jest.mock('./checksum');
 jest.mock('./fetch');
