+++ x-pack/plugins/screenshotting/server/config/schema.test.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { ConfigSchema } from './schema';
+import { ConfigSchema } from '@kbn/screenshotting-server';
 
 describe('ConfigSchema', () => {
   it(`should produce correct config for context {"dev": false,"dist": false}`, () => {
