+++ x-pack/plugins/screenshotting/server/screenshots/event_logger/index.test.ts
@@ -5,11 +5,11 @@
  * 2.0.
  */
 
-import moment from 'moment';
 import { loggingSystemMock } from '@kbn/core/server/mocks';
+import { ConfigType } from '@kbn/screenshotting-server';
+import moment from 'moment';
 import { Actions, EventLogger, ScreenshottingAction, Transactions } from '.';
 import { ElementPosition } from '../get_element_position_data';
-import { ConfigType } from '../../config';
 
 jest.mock('uuid', () => ({
   v4: () => 'NEW_UUID',
