+++ x-pack/plugins/screenshotting/tsconfig.json
@@ -16,17 +16,16 @@
     "@kbn/screenshot-mode-plugin",
     "@kbn/cloud-plugin",
     "@kbn/utility-types",
-    "@kbn/config-schema",
     "@kbn/logging",
     "@kbn/std",
     "@kbn/i18n",
     "@kbn/utils",
-    "@kbn/safer-lodash-set",
     "@kbn/core-logging-server-mocks",
     "@kbn/logging-mocks",
     "@kbn/core-http-server",
     "@kbn/core-plugins-server",
     "@kbn/task-manager-plugin",
+    "@kbn/screenshotting-server",
   ],
   "exclude": [
     "target/**/*",
